requirements.txt
opencv-python>=4.2.0
Pillow>=9.0.0
numpy>=1.20.0
tqdm>=4.60.0
pycocotools>=2.0.0
scikit-image>=0.19.0
PyYAML>=5.4.0
icrawler>=0.6.0